feat: oauth2 complete
This commit is to markdown the goals accomplished by the api. The OAuth2
part is complete, the API can create the basic URL to authenticate in
google API, remember that the main goal is only to create URLs but not send it.
The release include options to revoke, refresh, and exchange creation of
URLs that permits the connection between google.