refactor: module imports
When importing functions that do not belong to the original api was
showing. The purpose of this modification is to organize the code
in folders and decide where each function should belong to. All
methods and class that do not pertence to the api have now a leading
score, test for these file was also changed.