test: oauth2 I/O of tests
The results and expected results was added in a new folder. The reason
for that is because each input and expected results were too big to
maintain. The values are added in a json where "i" represents input
data in "fixture.json" and "o" represents output in "expected.json".
Each test unit has its own "i" and "o" that range from letters "a" to
"z" in each section that is being tested.